#294 Delete Teams
