+++ client/app/routes/dashboard/settings/organization/members.tsx
@@ -77,7 +77,7 @@ export default function Members() {
                       intent="secondary"
                       onClick={() => setShowAddMemberDialog(true)}
                     >
-                      Add Member
+                      Invite Member
                     </Button>
                   </th>
                 )}
