+++ client/app/routes/dashboard/settings/organization/teams/[teamId].tsx
@@ -4,10 +4,11 @@ import { useQuery } from "@tanstack/react-query";
 import { useTRPC } from "~/utils/trpc/react";
 import { useOrganizationContext } from "~/contexts/OrganizationContext";
 import Skeleton from "~/components/atoms/skeleton/Skeleton";
-import TeamMembers from "~/components/organisms/settings/TeamMembers";
-import { RepositoryAccess } from "~/components/organisms/settings/RepositoryAccess";
-import { ClusterAccess } from "~/components/organisms/settings/ClusterAccess";
+import TeamMembers from "../../../../../components/organisms/settings/organization/team/TeamMembers";
+import { RepositoryAccess } from "~/components/organisms/settings/organization/team/RepositoryAccess";
+import { ClusterAccess } from "~/components/organisms/settings/organization/team/ClusterAccess";
 import Breadcrumbs from "~/components/organisms/breadcrumbs/Breadcrumbs";
+import TeamDangerZone from "~/components/organisms/settings/organization/team/TeamDangerZone";
 
 export function loader() {
   return {
@@ -67,13 +68,15 @@ export default function TeamDetail() {
           </div>
 
           {organization.isOwner && (
-            <RepositoryAccess
-              teamId={Number(teamId)}
-              githubAppName={githubAppName}
-            />
+            <>
+              <RepositoryAccess
+                teamId={Number(teamId)}
+                githubAppName={githubAppName}
+              />
+              <ClusterAccess teamId={Number(teamId)} />
+              <TeamDangerZone teamId={Number(teamId)} />
+            </>
           )}
-
-          {organization.isOwner && <ClusterAccess teamId={Number(teamId)} />}
         </div>
       </div>
     </>
