+++ server/cmd/api/core/docs/core_core_swagger.json
@@ -1766,6 +1766,36 @@
                 }
             }
         },
+        "/teams/{teamId}": {
+            "delete": {
+                "tags": [
+                    "team"
+                ],
+                "summary": "Delete Team",
+                "operationId": "deleteTeam",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "User ID",
+                        "name": "X-User-ID",
+                        "in": "header",
+                        "required": true
+                    },
+                    {
+                        "type": "integer",
+                        "description": "Team ID",
+                        "name": "teamId",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK"
+                    }
+                }
+            }
+        },
         "/teams/{teamId}/clusters": {
             "get": {
                 "tags": [
