+++ server/cmd/api/core/docs/core_core_swagger.yaml
@@ -2073,6 +2073,26 @@ paths:
       summary: Toggle Project Preview Environment Enabled
       tags:
       - project
+  /teams/{teamId}:
+    delete:
+      operationId: deleteTeam
+      parameters:
+      - description: User ID
+        in: header
+        name: X-User-ID
+        required: true
+        type: string
+      - description: Team ID
+        in: path
+        name: teamId
+        required: true
+        type: integer
+      responses:
+        "200":
+          description: OK
+      summary: Delete Team
+      tags:
+      - team
   /teams/{teamId}/clusters:
     get:
       operationId: getTeamClusters
